pyproject.toml
[project]
name = "semana-da-fisica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "ignr>=2.2",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.3",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length=79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev semana_da_fisica/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=semana_da_fisica -vv'
post_test = 'coverage html'
